Move test-specific utils into test module
Update README + example
