+++ tests/runner/Cargo.toml
@@ -32,3 +32,4 @@ rand = "^0.8"
 md-5 = "^0.9"
 serde = "^1.0"
 serde_json = "^1.0"
+tokio = { version = "0.2.21", features = ["macros"] }
