+++ tests/runner/src/utils/lib.rs
@@ -4,7 +4,4 @@ pub mod tls;
 
 pub mod test_meta;
 
-pub mod smoke;
-pub mod concurrent;
-
 const VERSION: &str = include_str!("../../../../VERSION");
